requirements.txt
aiohttp==3.10.11
annotated-types==0.7.0
anyio==4.9.0
certifi==2025.7.14
distro==1.9.0
faiss-cpu==1.11.0.post1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
jiter==0.10.0
joblib==1.5.1
numpy==2.3.2
openai==1.97.1
packaging==25.0
psycopg2-binary==2.9.10
pydantic==2.11.7
pydantic_core==2.33.2
scikit-learn==1.7.1
scipy==1.16.1
sniffio==1.3.1
SQLAlchemy==2.0.41
threadpoolctl==3.6.0
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.1